requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
Werkzeug==3.0.3
psycopg2-binary==2.9.9

# Optional: Stripe payment integration
stripe==9.12.0

# Optional: production WSGI server
gunicorn==22.0.0